Added preview-image, modal, onHover effects for buttons.
Added preview-image, modal, onHover effects for buttons.

There is now a preview image which was missing on the github-pages site. There is also a hidden modal that appears when clicking the preview image to be able to see the apps functions.

Buttons now has an on hover effect when hovering on them.